+++ CHANGELOG.md
@@ -17,6 +17,7 @@ All notable changes to this project will be documented in this file.
   `DuckDB::Appender#append_blob`, `DuckDB::Appender#append_null`, `DuckDB::Appender#append_default`,
   `DuckDB::Appender#append_date`, `DuckDB::Appender#append_interval`, `DuckDB::Appender#append_time`,
   `DuckDB::Appender#append_timestamp`, `DuckDB::Appender#append_hugeint` failed.
+- add `DuckDB::PreparedStatement#bind_uhugeint`.
 
 # 1.2.0.0 - 2025-02-24
 - bump duckdb to 1.2.0.